+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/Repositories/AcademyRepository.cs
@@ -1,9 +1,9 @@
+using System.Globalization;
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Contexts;
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Extensions;
 using DfE.FindInformationAcademiesTrusts.Data.Repositories.Academy;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using System.Globalization;
 
 namespace DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Repositories;
 
@@ -57,8 +57,8 @@ public class AcademyRepository(IAcademiesDbContext academiesDbContext, ILogger<A
         var ofstedRatings =
             await academiesDbContext.MisEstablishments
                 .Where(me => urns.Contains(me.Urn!.Value))
-                .Select(me => new OfstedRatings(me.Urn!.Value, me.OverallEffectiveness, me.InspectionEndDate,
-                    me.PreviousFullInspectionOverallEffectiveness, me.PreviousInspectionEndDate))
+                .Select(me => new OfstedRatings(me.Urn!.Value, me.OverallEffectiveness, me.InspectionStartDate,
+                    me.PreviousFullInspectionOverallEffectiveness, me.PreviousInspectionStartDate))
                 .ToListAsync();
 
         // Look in MisFurtherEducationEstablishments for academies not found in MisEstablishments
@@ -133,32 +133,75 @@ public class AcademyRepository(IAcademiesDbContext academiesDbContext, ILogger<A
             .FirstOrDefaultAsync();
     }
 
+    public async Task<AcademyOverview[]> GetAcademiesInTrustOverviewAsync(string uid)
+    {
+        var academiesData = await academiesDbContext.GiasGroupLinks
+            .Where(gl => gl.GroupUid == uid)
+            .Join(
+                academiesDbContext.GiasEstablishments,
+                gl => gl.Urn!,
+                e => e.Urn.ToString(),
+                (gl, e) => new
+                {
+                    Urn = e.Urn.ToString(),
+                    LocalAuthority = e.LaName,
+                    NumberOfPupils = e.NumberOfPupils.ParseAsNullableInt(),
+                    SchoolCapacity = e.SchoolCapacity.ParseAsNullableInt()
+                })
+            .ToListAsync();
+
+        // Get URNs as ints
+        var urns = academiesData.Select(a => int.Parse(a.Urn)).ToArray();
+
+        // Fetch Ofsted ratings
+        var ofstedRatingsDict = await GetOfstedRatings(urns);
+
+        var academiesOverview = academiesData.Select(a =>
+        {
+            var currentOfstedRating = OfstedRatingScore.None;
+            if (ofstedRatingsDict.TryGetValue(a.Urn, out var ofstedRatings))
+            {
+                currentOfstedRating = ofstedRatings.Current?.OfstedRatingScore ?? OfstedRatingScore.None;
+            }
+
+            return new AcademyOverview(
+                a.Urn,
+                a.LocalAuthority ?? string.Empty,
+                a.NumberOfPupils,
+                a.SchoolCapacity,
+                currentOfstedRating
+            );
+        }).ToArray();
+
+        return academiesOverview;
+    }
+
     private sealed record OfstedRatings(int Urn, OfstedRating Current, OfstedRating Previous)
     {
-        public OfstedRatings(int urn, int? currentOverallEffectiveness, string? currentInspectionEndDate,
-            string? previousOverallEffectiveness, string? previousInspectionEndDate) :
+        public OfstedRatings(int urn, int? currentOverallEffectiveness, string? currentInspectionDate,
+            string? previousOverallEffectiveness, string? previousInspectionDate) :
             this(urn,
                 currentOverallEffectiveness,
-                currentInspectionEndDate,
+                currentInspectionDate,
                 previousOverallEffectiveness is null ? null : int.Parse(previousOverallEffectiveness),
-                previousInspectionEndDate)
+                previousInspectionDate)
         {
         }
 
-        public OfstedRatings(int urn, int? currentOverallEffectiveness, string? currentInspectionEndDate,
-            int? previousOverallEffectiveness, string? previousInspectionEndDate)
+        public OfstedRatings(int urn, int? currentOverallEffectiveness, string? currentInspectionDate,
+            int? previousOverallEffectiveness, string? previousInspectionDate)
             : this(urn,
                 currentOverallEffectiveness == null
                     ? OfstedRating.None
                     : new OfstedRating(
                         (OfstedRatingScore)currentOverallEffectiveness.Value,
-                        currentInspectionEndDate.ParseAsNullableDate()
+                        currentInspectionDate.ParseAsNullableDate()
                     ),
                 previousOverallEffectiveness == null
                     ? OfstedRating.None
                     : new OfstedRating(
                         (OfstedRatingScore)previousOverallEffectiveness,
-                        previousInspectionEndDate.ParseAsNullableDate()
+                        previousInspectionDate.ParseAsNullableDate()
                     )
             )
         {