+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Academy/AcademyOverview.cs
@@ -1,0 +1,9 @@
+﻿namespace DfE.FindInformationAcademiesTrusts.Data.Repositories.Academy;
+
+public record AcademyOverview(
+    string Urn,
+    string LocalAuthority,
+    int? NumberOfPupils,
+    int? SchoolCapacity,
+    OfstedRatingScore CurrentOfstedRating
+);