+++ DfE.FindInformationAcademiesTrusts.Data/Repositories/Academy/IAcademyRepository.cs
@@ -8,4 +8,5 @@ public interface IAcademyRepository
     Task<AcademyOfsted[]> GetAcademiesInTrustOfstedAsync(string uid);
     Task<AcademyPupilNumbers[]> GetAcademiesInTrustPupilNumbersAsync(string uid);
     Task<AcademyFreeSchoolMeals[]> GetAcademiesInTrustFreeSchoolMealsAsync(string uid);
+    Task<AcademyOverview[]> GetAcademiesInTrustOverviewAsync(string uid);
 }