+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Academies/AcademiesPageModel.cs
@@ -1,5 +1,6 @@
 ﻿using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Services;
+using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using Microsoft.AspNetCore.Mvc;
@@ -10,6 +11,7 @@ namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies
         ITrustProvider trustProvider,
         IDataSourceService dataSourceService,
         ITrustService trustService,
+        IAcademyService academyService,
         IExportService exportService,
         ILogger<AcademiesPageModel> logger,
         IDateTimeProvider dateTimeProvider
@@ -22,6 +24,7 @@ namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies
         {
             TrustSummaryServiceModel? trustSummary = await TrustService.GetTrustSummaryAsync(uid);
             Trust? allAcademiesDetails = await TrustProvider.GetTrustByUidAsync(uid);
+            AcademyOfstedServiceModel[] ofstedRatings = await academyService.GetAcademiesInTrustOfstedAsync(uid);
 
             if (trustSummary == null || allAcademiesDetails == null)
             {
@@ -31,7 +34,7 @@ namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies
             // Sanitize the trust name to remove any illegal characters
             string sanitizedTrustName = string.Concat(allAcademiesDetails.Name.Where(c => !Path.GetInvalidFileNameChars().Contains(c)));
 
-            var fileContents = ExportService.ExportAcademiesToSpreadsheetUsingProvider(allAcademiesDetails, trustSummary);
+            var fileContents = ExportService.ExportAcademiesToSpreadsheetUsingProvider(allAcademiesDetails, trustSummary, ofstedRatings);
             string fileName = $"{sanitizedTrustName}-{DateTimeProvider.Now:yyyy-MM-dd}.xlsx";
             string contentType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";