+++ tests/DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.UnitTests/Repositories/AcademyRepositoryTests.cs
@@ -118,11 +118,11 @@ public class AcademyRepositoryTests
 
         _mockAcademiesDbContext.AddMisEstablishments(new[]
         {
-            new MisEstablishment { Urn = urnsAsInt[0], OverallEffectiveness = 1, InspectionEndDate = "01/01/2022" },
-            new MisEstablishment { Urn = urnsAsInt[1], OverallEffectiveness = 2, InspectionEndDate = "29/02/2024" },
-            new MisEstablishment { Urn = urnsAsInt[2], OverallEffectiveness = 3, InspectionEndDate = "31/12/2022" },
-            new MisEstablishment { Urn = urnsAsInt[3], OverallEffectiveness = 4, InspectionEndDate = "15/10/2023" },
-            new MisEstablishment { Urn = urnsAsInt[4], OverallEffectiveness = null, InspectionEndDate = null }
+            new MisEstablishment { Urn = urnsAsInt[0], OverallEffectiveness = 1, InspectionStartDate = "01/01/2022" },
+            new MisEstablishment { Urn = urnsAsInt[1], OverallEffectiveness = 2, InspectionStartDate = "29/02/2024" },
+            new MisEstablishment { Urn = urnsAsInt[2], OverallEffectiveness = 3, InspectionStartDate = "31/12/2022" },
+            new MisEstablishment { Urn = urnsAsInt[3], OverallEffectiveness = 4, InspectionStartDate = "15/10/2023" },
+            new MisEstablishment { Urn = urnsAsInt[4], OverallEffectiveness = null, InspectionStartDate = null }
         });
         _mockAcademiesDbContext.AddMisFurtherEducationEstablishments(new[]
         {
@@ -165,11 +165,11 @@ public class AcademyRepositoryTests
 
         _mockAcademiesDbContext.AddMisEstablishments(new[]
         {
-            new MisEstablishment { Urn = urnsAsInt[0], PreviousFullInspectionOverallEffectiveness = "1", PreviousInspectionEndDate = "01/01/2022" },
-            new MisEstablishment { Urn = urnsAsInt[1], PreviousFullInspectionOverallEffectiveness = "2", PreviousInspectionEndDate = "29/02/2024" },
-            new MisEstablishment { Urn = urnsAsInt[2], PreviousFullInspectionOverallEffectiveness = "3", PreviousInspectionEndDate = "31/12/2022" },
-            new MisEstablishment { Urn = urnsAsInt[3], PreviousFullInspectionOverallEffectiveness = "4", PreviousInspectionEndDate = "15/10/2023" },
-            new MisEstablishment { Urn = urnsAsInt[4], PreviousFullInspectionOverallEffectiveness = null, PreviousInspectionEndDate = null }
+            new MisEstablishment { Urn = urnsAsInt[0], PreviousFullInspectionOverallEffectiveness = "1", PreviousInspectionStartDate = "01/01/2022" },
+            new MisEstablishment { Urn = urnsAsInt[1], PreviousFullInspectionOverallEffectiveness = "2", PreviousInspectionStartDate = "29/02/2024" },
+            new MisEstablishment { Urn = urnsAsInt[2], PreviousFullInspectionOverallEffectiveness = "3", PreviousInspectionStartDate = "31/12/2022" },
+            new MisEstablishment { Urn = urnsAsInt[3], PreviousFullInspectionOverallEffectiveness = "4", PreviousInspectionStartDate = "15/10/2023" },
+            new MisEstablishment { Urn = urnsAsInt[4], PreviousFullInspectionOverallEffectiveness = null, PreviousInspectionStartDate = null }
         });
         _mockAcademiesDbContext.AddMisFurtherEducationEstablishments(new[]
         {
@@ -232,7 +232,7 @@ public class AcademyRepositoryTests
             {
                 Urn = urns[0],
                 PreviousFullInspectionOverallEffectiveness = "1",
-                PreviousInspectionEndDate = "01/01/2022"
+                PreviousInspectionStartDate = "01/01/2022"
             }
         });
         _mockAcademiesDbContext.AddMisFurtherEducationEstablishments(new[]
@@ -397,6 +397,137 @@ public class AcademyRepositoryTests
         GetAcademiesInTrustFreeSchoolMealsAsync_should_return_empty_array_when_no_academies_linked_to_trust()
     {
         var result = await _sut.GetAcademiesInTrustFreeSchoolMealsAsync("1234");
+        result.Should().BeEmpty();
+    }
+
+    [Fact]
+    public async Task GetAcademiesInTrustOverviewAsync_should_return_academies_linked_to_trust()
+    {
+        // Arrange
+        var giasGroup = _mockAcademiesDbContext.AddGiasGroup("1234");
+        var giasEstablishments = Enumerable.Range(1000, 3).Select(n => new GiasEstablishment
+        {
+            Urn = n,
+            EstablishmentName = $"Academy {n}",
+            LaName = $"Local authority {n}",
+            NumberOfPupils = (n * 10).ToString(),
+            SchoolCapacity = (n * 15).ToString()
+        }).ToArray();
+
+        _mockAcademiesDbContext.AddGiasEstablishments(giasEstablishments);
+        _mockAcademiesDbContext.AddGiasGroupLinksForGiasEstablishmentsToGiasGroup(giasEstablishments, giasGroup);
+
+        // Act
+        var result = await _sut.GetAcademiesInTrustOverviewAsync("1234");
+
+        // Assert
+        result.Should().BeEquivalentTo(giasEstablishments,
+            options => options
+                .WithAutoConversion()
+                .ExcludingMissingMembers()
+                .WithMapping<AcademyOverview>(e => e.LaName, a => a.LocalAuthority)
+        );
+    }
+
+    // @formatter:max_line_length 500 - these tests are very difficult to read after line wrapping is enforced so increase max line length
+    [Fact]
+    public async Task GetAcademiesInTrustOverviewAsync_should_set_current_ofsted()
+    {
+        // Arrange
+        var giasGroup = _mockAcademiesDbContext.AddGiasGroup("1234");
+        var giasEstablishments = _mockAcademiesDbContext.AddGiasEstablishments(10);
+        _mockAcademiesDbContext.AddGiasGroupLinksForGiasEstablishmentsToGiasGroup(giasEstablishments, giasGroup);
+
+        var urnsAsInt = giasEstablishments.Select(gl => gl.Urn).ToArray();
+        var urns = urnsAsInt.Select(u => u.ToString()).ToArray();
+
+        _mockAcademiesDbContext.AddMisEstablishments(new[]
+        {
+            new MisEstablishment { Urn = urnsAsInt[0], OverallEffectiveness = 1, InspectionStartDate = "01/01/2022" },
+            new MisEstablishment { Urn = urnsAsInt[1], OverallEffectiveness = 2, InspectionStartDate = "29/02/2024" },
+            new MisEstablishment { Urn = urnsAsInt[2], OverallEffectiveness = 3, InspectionStartDate = "31/12/2022" },
+            new MisEstablishment { Urn = urnsAsInt[3], OverallEffectiveness = 4, InspectionStartDate = "15/10/2023" },
+            new MisEstablishment { Urn = urnsAsInt[4], OverallEffectiveness = null, InspectionStartDate = null }
+        });
+        _mockAcademiesDbContext.AddMisFurtherEducationEstablishments(new[]
+        {
+            new MisFurtherEducationEstablishment { ProviderUrn = urnsAsInt[5], OverallEffectiveness = 1, LastDayOfInspection = "01/01/2022" },
+            new MisFurtherEducationEstablishment { ProviderUrn = urnsAsInt[6], OverallEffectiveness = 2, LastDayOfInspection = "29/02/2024" },
+            new MisFurtherEducationEstablishment { ProviderUrn = urnsAsInt[7], OverallEffectiveness = 3, LastDayOfInspection = "31/12/2022" },
+            new MisFurtherEducationEstablishment { ProviderUrn = urnsAsInt[8], OverallEffectiveness = 4, LastDayOfInspection = "15/10/2023" },
+            new MisFurtherEducationEstablishment { ProviderUrn = urnsAsInt[9], OverallEffectiveness = null, LastDayOfInspection = null }
+        });
+
+        // Act
+        var result = await _sut.GetAcademiesInTrustOverviewAsync("1234");
+
+        // Assert
+        var dummyAcademyOverview = new AcademyOverview(string.Empty, string.Empty, null, null, OfstedRatingScore.None);
+
+        result.Should().BeEquivalentTo(new[]
+            {
+                dummyAcademyOverview with { Urn = urns[0], CurrentOfstedRating = OfstedRatingScore.Outstanding },
+                dummyAcademyOverview with { Urn = urns[1], CurrentOfstedRating = OfstedRatingScore.Good },
+                dummyAcademyOverview with { Urn = urns[2], CurrentOfstedRating = OfstedRatingScore.RequiresImprovement },
+                dummyAcademyOverview with { Urn = urns[3], CurrentOfstedRating = OfstedRatingScore.Inadequate },
+                dummyAcademyOverview with { Urn = urns[4], CurrentOfstedRating = OfstedRatingScore.None },
+                dummyAcademyOverview with { Urn = urns[5], CurrentOfstedRating = OfstedRatingScore.Outstanding },
+                dummyAcademyOverview with { Urn = urns[6], CurrentOfstedRating = OfstedRatingScore.Good },
+                dummyAcademyOverview with { Urn = urns[7], CurrentOfstedRating = OfstedRatingScore.RequiresImprovement },
+                dummyAcademyOverview with { Urn = urns[8], CurrentOfstedRating = OfstedRatingScore.Inadequate },
+                dummyAcademyOverview with { Urn = urns[9], CurrentOfstedRating = OfstedRatingScore.None }
+            },
+            options => options
+                .Including(a => a.Urn)
+                .Including(a => a.CurrentOfstedRating)
+        );
+    }
+    // @formatter:max_line_length restore
+
+    [Fact]
+    public async Task GetAcademiesInTrustOverviewAsync_should_handle_academies_with_missing_data()
+    {
+        var giasGroup = _mockAcademiesDbContext.AddGiasGroup("1234");
+        var giasEstablishment = new GiasEstablishment
+        {
+            Urn = 2000,
+            EstablishmentName = "Academy Missing Data",
+            LaName = null,
+            NumberOfPupils = null,
+            SchoolCapacity = null
+        };
+        _mockAcademiesDbContext.AddGiasEstablishments(new[] { giasEstablishment });
+        _mockAcademiesDbContext.AddGiasGroupLink(new GiasGroupLink
+        {
+            GroupUid = giasGroup.GroupUid,
+            Urn = giasEstablishment.Urn.ToString()
+        });
+
+        var result = await _sut.GetAcademiesInTrustOverviewAsync("1234");
+
+        result.Should().NotBeNull();
+        result.Length.Should().Be(1);
+
+        var academy = result.First();
+        academy.Urn.Should().Be("2000");
+        academy.LocalAuthority.Should().Be(string.Empty);
+        academy.NumberOfPupils.Should().BeNull();
+        academy.SchoolCapacity.Should().BeNull();
+    }
+
+    [Fact]
+    public async Task GetAcademiesInTrustOverviewAsync_should_return_empty_array_when_no_academies_linked_to_trust()
+    {
+        var result = await _sut.GetAcademiesInTrustOverviewAsync("1234");
+        result.Should().NotBeNull();
+        result.Should().BeEmpty();
+    }
+
+    [Fact]
+    public async Task GetAcademiesInTrustOverviewAsync_should_return_empty_array_when_trust_does_not_exist()
+    {
+        var result = await _sut.GetAcademiesInTrustOverviewAsync("non-existent-uid");
+        result.Should().NotBeNull();
         result.Should().BeEmpty();
     }
 }