+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/OverviewModelTests.cs
@@ -1,6 +1,5 @@
 ﻿using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
-using DfE.FindInformationAcademiesTrusts.Data.UnitTests.Mocks;
 using DfE.FindInformationAcademiesTrusts.Pages.Trusts;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using DfE.FindInformationAcademiesTrusts.UnitTests.Mocks;
@@ -10,23 +9,22 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts;
 
 public class OverviewModelTests
 {
-    private readonly Mock<ITrustProvider> _mockTrustProvider = new();
     private readonly OverviewModel _sut;
     private const string TrustUid = "1234";
     private readonly MockDataSourceService _mockDataSourceService = new();
-    private readonly Mock<ITrustService> _mockTrustRepository = new();
+    private readonly Mock<ITrustService> _mockTrustService = new();
+
+    private static readonly TrustOverviewServiceModel BaseTrustOverviewServiceModel =
+        new("1234", 0, new Dictionary<string, int>(), 0, 0, new Dictionary<OfstedRatingScore, int>());
 
     public OverviewModelTests()
     {
-        var dummyTrust = DummyTrustFactory.GetDummyTrust(TrustUid);
-        MockLogger<OverviewModel> logger = new();
+        _mockTrustService.Setup(t => t.GetTrustSummaryAsync(TrustUid))
+            .ReturnsAsync(new TrustSummaryServiceModel(TrustUid, "My Trust", "Multi-academy trust", 3));
 
-        _mockTrustProvider.Setup(tp => tp.GetTrustByUidAsync(TrustUid)).ReturnsAsync(dummyTrust);
-        _mockTrustRepository.Setup(t => t.GetTrustSummaryAsync(dummyTrust.Uid))
-            .ReturnsAsync(new TrustSummaryServiceModel(dummyTrust.Uid, dummyTrust.Name, dummyTrust.Type,
-                dummyTrust.Academies.Length));
-        _sut = new OverviewModel(_mockTrustProvider.Object, _mockDataSourceService.Object, logger.Object,
-                _mockTrustRepository.Object)
+        _sut = new OverviewModel(Mock.Of<ITrustProvider>(), _mockDataSourceService.Object,
+                new MockLogger<OverviewModel>().Object,
+                _mockTrustService.Object)
             { Uid = TrustUid };
     }
 
@@ -36,203 +34,147 @@ public class OverviewModelTests
         _sut.PageName.Should().Be("Overview");
     }
 
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_numbers_on_multi_academy_trust()
+    [Theory]
+    [InlineData(0)]
+    [InlineData(200)]
+    [InlineData(1125)]
+    public async Task OnGetAsync_sets_pupil_numbers(int num)
     {
-        SetupTrustWithMultipleAcademies();
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid))
+            .ReturnsAsync(BaseTrustOverviewServiceModel with { TotalPupilNumbers = num });
+
         await _sut.OnGetAsync();
-        _sut.TotalPupilNumbersInTrust.Should().Be(1125);
+
+        _sut.TotalPupilNumbers.Should().Be(num);
     }
 
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_numbers_on_single_academy_trust()
+    [Theory]
+    [InlineData(0)]
+    [InlineData(200)]
+    [InlineData(1125)]
+    public async Task OnGetAsync_sets_pupil_capacity(int num)
     {
-        SetupTrustWithSingleAcademy();
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid))
+            .ReturnsAsync(BaseTrustOverviewServiceModel with { TotalCapacity = num });
+
         await _sut.OnGetAsync();
-        _sut.TotalPupilNumbersInTrust.Should().Be(200);
+
+        _sut.TotalCapacity.Should().Be(num);
     }
 
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_numbers_on_trust_with_no_academies()
+    [Theory]
+    [InlineData(400, 300)]
+    [InlineData(0, 0)]
+    public async Task OnGetAsync_sets_pupil_percentage(int totalCapacity, int totalPupilNumbers)
     {
-        SetupTrustWithNoAcademies();
-        await _sut.OnGetAsync();
-        _sut.TotalPupilNumbersInTrust.Should().Be(0);
-    }
+        var serviceModel =
+            BaseTrustOverviewServiceModel with
+            {
+                TotalCapacity = totalCapacity, TotalPupilNumbers = totalPupilNumbers
+            };
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid)).ReturnsAsync(serviceModel);
 
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_capacity_on_multi_academy_trust()
-    {
-        SetupTrustWithMultipleAcademies();
         await _sut.OnGetAsync();
-        _sut.TotalPupilCapacityInTrust.Should().Be(1950);
-    }
 
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_capacity_on_single_academy_trust()
-    {
-        SetupTrustWithSingleAcademy();
-        await _sut.OnGetAsync();
-        _sut.TotalPupilCapacityInTrust.Should().Be(250);
-    }
-
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_capacity_on_trust_with_no_academies()
-    {
-        SetupTrustWithNoAcademies();
-        await _sut.OnGetAsync();
-        _sut.TotalPupilCapacityInTrust.Should().Be(0);
-    }
-
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_percentage_on_multi_academy_trust()
-    {
-        SetupTrustWithMultipleAcademies();
-        await _sut.OnGetAsync();
-        _sut.TotalPercentageCapacityInTrust.Should().Be(58);
-    }
-
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_percentage_on_single_academy_trust()
-    {
-        SetupTrustWithSingleAcademy();
-        await _sut.OnGetAsync();
-        _sut.TotalPercentageCapacityInTrust.Should().Be(80);
-    }
-
-
-    [Fact]
-    public async Task OnGetAsync_sets_pupil_percentage_on_trust_with_no_academies()
-    {
-        SetupTrustWithNoAcademies();
-        await _sut.OnGetAsync();
-        _sut.TotalPercentageCapacityInTrust.Should().BeNull();
+        _sut.PercentageFull.Should().Be(serviceModel.PercentageFull);
     }
 
     [Fact]
     public async Task OnGetAsync_sets_list_of_local_authorities()
     {
-        SetupTrustWithMultipleAcademies();
-        var expectedLocalAuthorityCount = new (string? Authority, int Total)[]
+        var overviewWithLocalAuthorities = BaseTrustOverviewServiceModel with
         {
-            ("localAuth1", 6),
-            ("localAuth2", 1)
+            AcademiesByLocalAuthority = new Dictionary<string, int>
+            {
+                { "localAuth1", 6 },
+                { "localAuth2", 1 }
+            }
         };
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid)).ReturnsAsync(overviewWithLocalAuthorities);
 
         await _sut.OnGetAsync();
-        var result = _sut.AcademiesInEachLocalAuthority;
-        result.Should().BeEquivalentTo(expectedLocalAuthorityCount);
+
+        _sut.AcademiesInEachLocalAuthority
+            .Should()
+            .BeEquivalentTo(new (string Authority, int Total)[]
+            {
+                ("localAuth1", 6),
+                ("localAuth2", 1)
+            });
     }
 
-    [Fact]
-    public async Task OnGetAsync_sets_ofsted_ratings_list_correctly()
+    [Theory]
+    [InlineData(OfstedRatingScore.Outstanding, 3)]
+    [InlineData(OfstedRatingScore.Good, 4)]
+    [InlineData(OfstedRatingScore.Inadequate, 5)]
+    [InlineData(OfstedRatingScore.RequiresImprovement, 6)]
+    [InlineData(OfstedRatingScore.None, 7)]
+    public async Task GetNumberOfAcademiesWithOfstedRating_returns_correct_number_of_ratings(OfstedRatingScore score,
+        int expectedCount)
     {
-        SetupTrustWithMultipleAcademies();
-        var expectedOfstedRatingCount = new (OfstedRatingScore Rating, int Total)[]
+        var overviewWithOfstedRatings = BaseTrustOverviewServiceModel with
         {
-            (OfstedRatingScore.Outstanding, 3),
-            (OfstedRatingScore.Good, 1),
-            (OfstedRatingScore.Inadequate, 1),
-            (OfstedRatingScore.RequiresImprovement, 1),
-            (OfstedRatingScore.None, 1)
+            OfstedRatings = new Dictionary<OfstedRatingScore, int>
+            {
+                { OfstedRatingScore.Outstanding, 3 },
+                { OfstedRatingScore.Good, 4 },
+                { OfstedRatingScore.Inadequate, 5 },
+                { OfstedRatingScore.RequiresImprovement, 6 },
+                { OfstedRatingScore.None, 7 }
+            }
         };
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid)).ReturnsAsync(overviewWithOfstedRatings);
 
         await _sut.OnGetAsync();
-        var result = _sut.OfstedRatings;
-        result.Should().BeEquivalentTo(expectedOfstedRatingCount);
+
+        _sut.GetNumberOfAcademiesWithOfstedRating(score).Should().Be(expectedCount);
     }
 
-    [Fact]
-    public async Task OnGetAsync_ofsted_score_count_returns_correct_number_of_ratings()
+    [Theory]
+    [InlineData(OfstedRatingScore.Outstanding)]
+    [InlineData(OfstedRatingScore.Good)]
+    [InlineData(OfstedRatingScore.Inadequate)]
+    [InlineData(OfstedRatingScore.RequiresImprovement)]
+    [InlineData(OfstedRatingScore.None)]
+    public async Task GetNumberOfAcademiesWithOfstedRating_returns_zero_ratings_when_non_exist(
+        OfstedRatingScore missingRating)
     {
-        SetupTrustWithMultipleAcademies();
+        var ofstedRatings = new Dictionary<OfstedRatingScore, int>
+        {
+            { OfstedRatingScore.Outstanding, 3 },
+            { OfstedRatingScore.Good, 4 },
+            { OfstedRatingScore.Inadequate, 5 },
+            { OfstedRatingScore.RequiresImprovement, 6 },
+            { OfstedRatingScore.None, 7 }
+        };
+        ofstedRatings.Remove(missingRating);
+
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid))
+            .ReturnsAsync(BaseTrustOverviewServiceModel with { OfstedRatings = ofstedRatings });
+
         await _sut.OnGetAsync();
-        var result = _sut.GetNumberOfAcademiesWithOfstedRating(OfstedRatingScore.Outstanding);
-        result.Should().Be(3);
+
+        _sut.GetNumberOfAcademiesWithOfstedRating(missingRating).Should().Be(0);
     }
 
-    [Fact]
-    public async Task OnGetAsync_ofsted_score_count_returns_zero_ratings_when_non_exist()
+    [Theory]
+    [InlineData(0)]
+    [InlineData(1)]
+    [InlineData(90)]
+    public async Task OnGetAsync_returns_correct_number_of_academies(int num)
     {
-        SetupTrustWithSingleAcademy();
+        _mockTrustService.Setup(t => t.GetTrustOverviewAsync(TrustUid))
+            .ReturnsAsync(BaseTrustOverviewServiceModel with { TotalAcademies = num });
+
         await _sut.OnGetAsync();
-        var result = _sut.GetNumberOfAcademiesWithOfstedRating(OfstedRatingScore.Good);
-        result.Should().Be(0);
-    }
 
-    [Fact]
-    public async Task OnGetAsync_returns_correct_number_of_academies_for_single_academy_trust()
-    {
-        SetupTrustWithSingleAcademy();
-        await _sut.OnGetAsync();
-        var result = _sut.NumberOfAcademiesInTrust;
-        result.Should().Be(1);
-    }
-
-    [Fact]
-    public async Task OnGetAsync_returns_correct_number_of_academies_for_multi_academy_trust()
-    {
-        SetupTrustWithMultipleAcademies();
-        await _sut.OnGetAsync();
-        var result = _sut.NumberOfAcademiesInTrust;
-        result.Should().Be(7);
-    }
-
-    [Fact]
-    public async Task OnGetAsync_returns_correct_number_of_academies_for_trust_with_no_academies()
-    {
-        SetupTrustWithNoAcademies();
-        await _sut.OnGetAsync();
-        var result = _sut.NumberOfAcademiesInTrust;
-        result.Should().Be(0);
-    }
-
-    private static Academy DummyAcademy(int urn, string localAuthority, int numberOfPupils,
-        int schoolCapacity,
-        OfstedRatingScore currentOfstedRatingScore)
-    {
-        return DummyAcademyFactory.GetDummyAcademy(
-            urn,
-            localAuthority: localAuthority,
-            numberOfPupils: numberOfPupils,
-            schoolCapacity: schoolCapacity,
-            currentOfstedRatingScore: currentOfstedRatingScore);
-    }
-
-    private void SetupTrustWithMultipleAcademies()
-    {
-        _mockTrustProvider.Setup(tp => tp.GetTrustByUidAsync(TrustUid))
-            .ReturnsAsync(DummyTrustFactory.GetDummyTrust(TrustUid, academies: new[]
-            {
-                DummyAcademy(1, "localAuth1", 50, 150, OfstedRatingScore.Outstanding),
-                DummyAcademy(2, "localAuth1", 100, 200, OfstedRatingScore.Good),
-                DummyAcademy(3, "localAuth1", 150, 250, OfstedRatingScore.RequiresImprovement),
-                DummyAcademy(4, "localAuth1", 200, 300, OfstedRatingScore.Inadequate),
-                DummyAcademy(5, "localAuth1", 0, 500, OfstedRatingScore.None),
-                DummyAcademy(6, "localAuth1", 600, 500, OfstedRatingScore.Outstanding),
-                DummyAcademy(7, "localAuth2", 25, 50, OfstedRatingScore.Outstanding)
-            }));
-    }
-
-    private void SetupTrustWithSingleAcademy()
-    {
-        _mockTrustProvider.Setup(tp => tp.GetTrustByUidAsync(TrustUid))
-            .ReturnsAsync(DummyTrustFactory.GetDummyTrust(TrustUid, academies: new[]
-            {
-                DummyAcademy(42, "localAuth1", 200, 250, OfstedRatingScore.Outstanding)
-            }));
-    }
-
-    private void SetupTrustWithNoAcademies()
-    {
-        _mockTrustProvider.Setup(tp => tp.GetTrustByUidAsync(TrustUid))
-            .ReturnsAsync(DummyTrustFactory.GetDummyTrust(TrustUid, academies: null));
+        _sut.TotalAcademies.Should().Be(num);
     }
 
     [Fact]
     public async Task OnGetAsync_returns_NotFoundResult_if_Trust_is_not_found()
     {
-        _mockTrustRepository.Setup(t => t.GetTrustSummaryAsync(TrustUid)).ReturnsAsync((TrustSummaryServiceModel?)null);
+        _mockTrustService.Setup(t => t.GetTrustSummaryAsync(TrustUid)).ReturnsAsync((TrustSummaryServiceModel?)null);
         var result = await _sut.OnGetAsync();
         result.Should().BeOfType<NotFoundResult>();
     }